refactor: move JoinMarket client log parsing into its own module
Move joinmarket_parse_coinjoin_logs into joinmarket_client so the JoinMarket
branch has a module of its own and parse_cj_logs keeps the shared parsing core.

Resolve extract_tx_info and SM lazily through parse_cj_logs to avoid an import
cycle and preserve test patching. Keep parsed records unchanged while
standardizing progress and decoding-failure logging.
